+++ Pandora.Application/ViewModel/WalletViewModel.cs
@@ -3,8 +3,8 @@
     public class WalletViewModel
     {
         public int WalletType { get; set; }
-        public decimal Balance { get; set; }
-        public decimal InvestedBalance { get; set; }
-        public decimal AvailableBalance { get; set; }
+        public double Balance { get; set; }
+        public double InvestedBalance { get; set; }
+        public double AvailableBalance { get; set; }
     }
 }
