+++ Pandora.Domain/Domain/Plan.cs
@@ -6,8 +6,8 @@ namespace Pandora.Domain.Domain
     public class Plan : BaseEntity
     {
         public string Name { get; set; }
-        public double ProfitPercent { get; set; }
-        public double ReferralPercent { get; set; }
+        public int ProfitPercent { get; set; }
+        public int ReferralPercent { get; set; }
         public int MinimumDeposit { get; set; }
         public int Duration { get; set; }
